Separate extensions for UIKit and AppKit
