+++ Sources/DesignResourcesKit/Colors/ColorPalletes/ColorPaletteDefinition+AppKit.swift
@@ -1,0 +1,33 @@
+//
+//  ColorPaletteDefinition+AppKit.swift
+//
+//  Copyright © 2025 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import SwiftUI
+
+#if canImport(AppKit)
+
+extension ColorPaletteDefinition {
+    static func nsColor(for designSystemColor: DesignSystemColor) -> NSColor {
+        dynamicColor(for: designSystemColor).nsColor
+    }
+
+    static func nsColor(for singleUseColor: SingleUserColor) -> NSColor {
+        dynamicColor(for: singleUseColor).nsColor
+    }
+}
+
+#endif
