+++ Sources/DesignResourcesKit/Colors/ColorPalletes/ColorPaletteDefinition+UIKit.swift
@@ -1,0 +1,33 @@
+//
+//  ColorPaletteDefinition+UIKit.swift
+//
+//  Copyright © 2025 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import SwiftUI
+
+#if canImport(UIKit)
+
+extension ColorPaletteDefinition {
+    static func uiColor(for designSystemColor: DesignSystemColor) -> UIColor {
+        dynamicColor(for: designSystemColor).uiColor
+    }
+
+    static func uiColor(for singleUseColor: SingleUserColor) -> UIColor {
+        dynamicColor(for: singleUseColor).uiColor
+    }
+}
+
+#endif
